% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_pronoun.R
\name{check_data_pronoun}
\alias{check_data_pronoun}
\title{Check possible notes/warnings while using .data$ pronoun and NSE/SE
evaluation}
\usage{
check_data_pronoun(df)
}
\arguments{
\item{df}{a tibble data.frame with at least two columns: \code{col_a} and \code{col_b}}
}
\value{
A tibble data.frame
}
\description{
Check possible notes/warnings while using .data$ pronoun and NSE/SE
evaluation
}
