% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy_df}
\alias{dummy_df}
\title{A dummy data.frame}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
dummy_df
}
\description{
A dummy data.frame
}
\examples{
\dontrun{
# mica.rda was created with the code below.
dummy_df <- dplyr::tibble(
  col_a = c(1, 2, 3),
  col_b = c(4, 5, 6),
  col_c = c(7, 8, 9)
)
save(dummy_df, file = "data/dummy_df.rda")
}
}
\keyword{datasets}
