% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyR-package.R
\docType{package}
\name{dummyR-package}
\alias{dummyR}
\alias{dummyR-package}
\title{dummyR: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Tiago Maié \email{tiagomaie+Rdev@example.com}

}
\keyword{internal}
