% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyR.R
\name{dummyR}
\alias{dummyR}
\alias{dummyr}
\alias{dumr}
\title{A dummy package to show how to easily create a package}
\usage{
dummyR(input_data)

dummyr(input_data)

dumr(input_data)
}
\arguments{
\item{input_data}{Parameter description.
A longer parameter description.}
}
\value{
Description of what the function returns.
}
\description{
Some dummy functions
}
\examples{
# Example of how to use this function
library('dummyR')

# load data
data(dataset_B)

# run dummyR
dummyR(input_data = dataset_B)

}
