% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Interactive Query}
\usage{
load_data(storage = NA, path = NA)
}
\arguments{
\item{storage}{type of data storage user wishes to query}

\item{path}{path where data resides}
}
\description{
Walking user through loading of data from multiple sources
}
