% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tidyselect.R
\name{check_tidyselect}
\alias{check_tidyselect}
\title{Check possible notes/warnings while using .data$ pronoun and NSE/SE
evaluation, NSE/SE}
\usage{
check_tidyselect(df, colToChange)
}
\arguments{
\item{df}{a tibble data.frame with at least two columns: \code{col_a} and \code{col_b}}

\item{colToChange}{character with column name to change}
}
\value{
A tibble data.frame
}
\description{
Check possible notes/warnings while using .data$ pronoun and NSE/SE
evaluation, NSE/SE
}
\examples{
library(dummyr)
check_tidyselect(dummy_df, "col_c")
}
